(chore): move server tests into __tests__/server directory
