+++ __tests__/server/controllers/roles.spec.js
@@ -1,8 +1,8 @@
 import { expect } from 'chai';
 import supertest from 'supertest';
-import app from '../../src/server';
-import { hashPassword } from '../../server/helpers/helper';
-import models from '../../server/models';
+import app from '../../../src/server';
+import { hashPassword } from '../../../server/helpers/helper';
+import models from '../../../server/models';
 
 const { Role, User } = models;
 const request = supertest.agent(app);
