+++ __tests__/server/controllers/search.spec.js
@@ -1,7 +1,7 @@
 import { expect } from 'chai';
 import supertest from 'supertest';
-import app from '../../src/server';
-import models from '../../server/models';
+import app from '../../../src/server';
+import models from '../../../server/models';
 
 const { Document, Role, User } = models;
 const request = supertest.agent(app);
